+++ Final_Project/Final_Project.vcxproj
@@ -150,17 +150,14 @@
     <ClCompile Include="archiveManager.c" />
     <ClCompile Include="buyAnApt.c" />
     <ClCompile Include="commander.c" />
+    <ClCompile Include="DataLoader.c" />
     <ClCompile Include="freeData.c" />
     <ClCompile Include="getAnApt.c" />
     <ClCompile Include="main.c" />
-    <ClCompile Include="ReadDataFromFiles.c" />
-    <ClCompile Include="test_main.c" />
-    <ClCompile Include="WriteDataToFiles.c" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="apartment.h" />
-    <ClInclude Include="DataLoader.h" />
-    <ClInclude Include="history_structsAndConsts.h" />
+    <ClInclude Include="dataLoader_header.h" />
+    <ClInclude Include="structsAndDefinitions.h" />
     <ClInclude Include="declerations.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
