+++ Final_Project/history.txt
@@ -1,48 +1,13 @@
-get-an-apt -MinimumNumRooms 0
+get-an-apt -Date 01102016 -sr
+get-an-apt -MinimumNumRooms 4 -sr
+get-an-apt -MinimumNumRooms 4 -sr
+add-an-apt "David Ben Gurion 25 Herzliya" 2200000 5 01 08 16
+get-an-apt -MinimumNumRooms 3 -MaximumNumRooms 6 -MaximumPrice 1760000 -s
+buy-an-apt 2
+get-an-apt -MaximumNumRooms 5 -s
+get-an-apt -MinimumNumRooms 4 -sr
+get-an-apt -MinimumNumRooms 3 -MaximumNumRooms 5 -MaximumPrice 1750000 -s
 add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
+add-an-apt "Ben Gurion 25 Herzliya" 2200000 5 01 08 16
 add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-get-an-apt -MinimumNumRooms 0
-delete-an-apt -Enter 1
-get-an-apt -MinimumNumRooms 0
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-delete-an-apt 1
-get-an-apt -MinimumNumRooms 0
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-get-an-apt -MinimumNumRooms 0
-delete-an-apt -Enter 2
-get-an-apt -MinimumNumRooms 0
-buy-an-apt 1
-get-an-apt -MinimumNumRooms 0
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-add-an-apt "Dizengoff 180 Tel Aviv" 2000000 4 28 04 16
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Dizengoff 180 Tel Aviv" 2000000 4 28 04 16
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-delete-an-apt -Enter 1
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-get-an-apt -MinimumNumRooms 0
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-get-an-apt -MinimumNumRooms 0
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-get-an-apt -MinimumNumRooms 0
-buy-an-apt 1
-delete-an-apt -Enter 1
-get-an-apt -MinimumNumRooms 0
-add-an-apt "Rothschild 67 Rishon Le Zion" 1700000 3 03 10 16
-add-an-apt "Ben Gurion 25 Herzliya" 200000 5 01 08 16
-add-an-apt "Gordon 85 Holon" 1500000 3 15 06 17
-get-an-apt -MinimumNumRooms 0
+add-an-apt "Dizengoff 180 Tel Aviv" 2000000 4 28 09 16
